Benchmarks:

goos: linux
goarch: amd64
pkg: github.com/spmadness/otus-go-hw/hw10_program_optimization
cpu: AMD Ryzen 7 5800H with Radeon Graphics

                 │ benchmark_results/old.txt │      benchmark_results/new.txt      │
                 │          sec/op           │   sec/op     vs base                │

  GetUsers-16                     1533.0µ ± 2%   229.8µ ± 1%  -85.01% (p=0.000 n=10)
CountDomains-16                146.844m ± 2%   6.932m ± 2%  -95.28% (p=0.000 n=10)
GetDomainStat-16                2090.5µ ± 6%   234.4µ ± 3%  -88.79% (p=0.000 n=10)
geomean                          7.778m        720.0µ       -90.74%

                 │ benchmark_results/old.txt │      benchmark_results/new.txt       │
                 │           B/op            │     B/op      vs base                │

GetUsers-16                   247.340Ki ± 2%   9.076Ki ± 1%  -96.33% (p=0.000 n=10)
CountDomains-16              136512.4Ki ± 0%   546.6Ki ± 0%  -99.60% (p=0.000 n=10)
GetDomainStat-16              649.453Ki ± 9%   9.450Ki ± 3%  -98.54% (p=0.000 n=10)
geomean                         2.733Mi        36.06Ki       -98.71%

                 │ benchmark_results/old.txt │      benchmark_results/new.txt      │
                 │         allocs/op         │  allocs/op   vs base                │

GetUsers-16                      1790.5 ± 2%    121.5 ± 3%  -93.21% (p=0.000 n=10)
CountDomains-16                1844.38k ± 0%   22.21k ± 0%  -98.80% (p=0.000 n=10)
GetDomainStat-16                 6458.5 ± 9%    131.5 ± 6%  -97.96% (p=0.000 n=10)
geomean                          27.73k         708.0       -97.45%